log match stats to google analytics
