+++ packages/desktop/hooks/LocalCombatsContext/index.tsx
@@ -1,14 +1,14 @@
-import { IArenaMatch, IShuffleRound } from '@wowarenalogs/parser';
+import { AtomicArenaCombat, buildQueryHelpers, CombatUnitSpec, CombatUnitType } from '@wowarenalogs/parser';
 import { logAnalyticsEvent, uploadCombatAsync, useAuth } from '@wowarenalogs/shared';
+import _ from 'lodash';
+import moment from 'moment';
 import React, { useContext, useEffect, useState } from 'react';
 
 import { useAppConfig } from '../AppConfigContext';
 
-type ParserCombatData = IArenaMatch | IShuffleRound;
-
 interface ILocalCombatsContextData {
-  localCombats: ParserCombatData[];
-  appendCombat: (combat: ParserCombatData) => void;
+  localCombats: AtomicArenaCombat[];
+  appendCombat: (combat: AtomicArenaCombat) => void;
 }
 
 const LocalCombatsContext = React.createContext<ILocalCombatsContextData>({
@@ -22,8 +22,85 @@ interface IProps {
   children: React.ReactNode | React.ReactNode[];
 }
 
+const logCombatAnalyticsAsync = async (combat: AtomicArenaCombat) => {
+  // const isPackaged = await window.wowarenalogs.app?.getIsPackaged();
+  // if (!isPackaged) {
+  //   return;
+  // }
+
+  const averageMMR =
+    combat.dataType === 'ArenaMatch'
+      ? ((combat.endInfo?.team0MMR || 0) + (combat.endInfo?.team1MMR || 0)) / 2
+      : ((combat.shuffleMatchEndInfo?.team0MMR || 0) + (combat.shuffleMatchEndInfo?.team1MMR || 0)) / 2;
+  const unitsList = _.values(combat.units).map((c) => ({
+    id: c.id,
+    name: c.name,
+    info: c.info,
+    type: c.type,
+    class: c.class,
+    spec: c.spec,
+    reaction: c.reaction,
+  }));
+  const players = unitsList.filter((u) => u.type === CombatUnitType.Player);
+  const team0specs = players
+    .filter((u) => u.info?.teamId === '0')
+    .map((u) => (u.spec === CombatUnitSpec.None ? `c${u.class}` : u.spec))
+    .sort()
+    .join('_');
+  const team1specs = players
+    .filter((u) => u.info?.teamId === '1')
+    .map((u) => (u.spec === CombatUnitSpec.None ? `c${u.class}` : u.spec))
+    .sort()
+    .join('_');
+  const indices = buildQueryHelpers(combat, true);
+
+  const commonProperties = {
+    wowVersion: combat.wowVersion,
+    combatId: combat.id,
+    date: moment(combat.startTime).format('YYYY-MM-DD'),
+    bracket: combat.startInfo.bracket,
+    zoneId: combat.startInfo.zoneId,
+    durationInSeconds: combat.durationInSeconds,
+    averageMMR,
+  };
+
+  // google analytics limitations:
+  // - event names can be up to 40 characters long
+  // - property names must be 40 characters or less
+  // - property values must be 100 characters or less
+  // - less than 25 properties
+  logAnalyticsEvent('event_NewMatchProcessed', {
+    ...commonProperties,
+    winningTeamSpecs: combat.winningTeamId === '0' ? team0specs : team1specs,
+    losingTeamSpecs: combat.winningTeamId === '1' ? team0specs : team1specs,
+    singleSidedSpecIndices: `|${indices.singleSidedSpecs.join('|')}|`,
+  });
+
+  logAnalyticsEvent('event_NewCompRecord', {
+    ...commonProperties,
+    specs: combat.winningTeamId === '0' ? team0specs : team1specs,
+    result: 'win',
+  });
+  logAnalyticsEvent('event_NewCompRecord', {
+    ...commonProperties,
+    specs: combat.winningTeamId === '1' ? team0specs : team1specs,
+    result: 'lose',
+  });
+
+  players.forEach((p) => {
+    logAnalyticsEvent('event_NewPlayerRecord', {
+      ...commonProperties,
+      name: p.name,
+      rating: p.info?.personalRating ?? 0,
+      highestRating: p.info?.highestPvpTier ?? 0,
+      spec: p.spec,
+      result: p.info?.teamId === combat.winningTeamId ? 'win' : 'lose',
+    });
+  });
+};
+
 export const LocalCombatsContextProvider = (props: IProps) => {
-  const [combats, setCombats] = useState<ParserCombatData[]>([]);
+  const [combats, setCombats] = useState<AtomicArenaCombat[]>([]);
   const auth = useAuth();
   const { wowInstallations } = useAppConfig();
 
@@ -33,25 +110,19 @@ export const LocalCombatsContextProvider = (props: IProps) => {
       window.wowarenalogs.logs?.startLogWatcher(wowDirectory, wowVersion);
 
       window.wowarenalogs.logs?.handleNewCombat((_event, combat) => {
-        if (
-          window.wowarenalogs.app?.getIsPackaged().then((isPackaged) => {
-            if (isPackaged) {
-              logAnalyticsEvent('event_NewMatchProcessed', {
-                wowVersion: combat.wowVersion,
-              });
-            }
-          })
-        )
-          if (wowVersion === combat.wowVersion) {
-            uploadCombatAsync(combat, auth.battlenetId);
-            setCombats((prev) => {
-              return prev.concat([combat]);
-            });
-          }
+        if (wowVersion === combat.wowVersion) {
+          uploadCombatAsync(combat, auth.battlenetId);
+          logCombatAnalyticsAsync(combat);
+
+          setCombats((prev) => {
+            return prev.concat([combat]);
+          });
+        }
       });
 
       window.wowarenalogs.logs?.handleSoloShuffleRoundEnded((_event, combat) => {
         if (wowVersion === combat.wowVersion) {
+          logCombatAnalyticsAsync(combat);
           setCombats((prev) => {
             return prev.concat([combat]);
           });
